fixes & warnings

+++ assembler/src/debruijn/path_extend/bidirectional_path.hpp
@@ -854,10 +854,10 @@ private:
 
 };
 
-int FirstNotEqualPosition(const BidirectionalPath& path1, int pos1,
-                       const BidirectionalPath& path2, int pos2) {
-    int cur_pos1 = pos1;
-    int cur_pos2 = pos2;
+int FirstNotEqualPosition(const BidirectionalPath& path1, size_t pos1,
+                       const BidirectionalPath& path2, size_t pos2) {
+    int cur_pos1 = (int) pos1;
+    int cur_pos2 = (int) pos2;
     while (cur_pos1 >= 0 && cur_pos2 >= 0) {
         if (path1.At(cur_pos1) == path2.At(cur_pos2)) {
             cur_pos1--;
@@ -868,15 +868,16 @@ int FirstNotEqualPosition(const BidirectionalPath& path1, int pos1,
     }
     return -1;
 }
-bool EqualBegins(const BidirectionalPath& path1, int pos1,
-                 const BidirectionalPath& path2, int pos2) {
+bool EqualBegins(const BidirectionalPath& path1, size_t pos1,
+                 const BidirectionalPath& path2, size_t pos2) {
     return FirstNotEqualPosition(path1, pos1, path2, pos2) == -1;
 }
-int LastNotEqualPosition(const BidirectionalPath& path1, int pos1,
-                      const BidirectionalPath& path2, int pos2) {
-    int cur_pos1 = pos1;
-    int cur_pos2 = pos2;
-    while (cur_pos1 < (int) path1.Size() && cur_pos2 < (int) path2.Size()) {
+
+size_t LastNotEqualPosition(const BidirectionalPath& path1, size_t pos1,
+                      const BidirectionalPath& path2, size_t pos2) {
+    size_t cur_pos1 = pos1;
+    size_t cur_pos2 = pos2;
+    while (cur_pos1 < path1.Size() && cur_pos2 < path2.Size()) {
         if (path1.At(cur_pos1) == path2.At(cur_pos2)) {
             cur_pos1++;
             cur_pos2++;
@@ -884,11 +885,11 @@ int LastNotEqualPosition(const BidirectionalPath& path1, int pos1,
             return cur_pos1;
         }
     }
-    return -1;
+    return -1UL;
 }
-bool EqualEnds(const BidirectionalPath& path1, int pos1,
-               const BidirectionalPath& path2, int pos2) {
-    return LastNotEqualPosition(path1, pos1, path2, pos2) == -1;
+bool EqualEnds(const BidirectionalPath& path1, size_t pos1,
+               const BidirectionalPath& path2, size_t pos2) {
+    return LastNotEqualPosition(path1, pos1, path2, pos2) == -1UL;
 }
 bool PathIdCompare(const BidirectionalPath* p1, const BidirectionalPath* p2) {
     return p1->GetId() < p2->GetId();
@@ -1130,8 +1131,7 @@ size_t LoopDetector::LoopEdges(size_t skip_identical_edges, size_t min_cycle_app
 bool LoopDetector::PathIsLoop(size_t edges) const {
     for (size_t i = 0; i < edges; ++i) {
         EdgeId e = path_->At(i);
-        for (int j = (int) path_->Size() - (edges - (int) i); j >= 0; j -=
-                (int) edges) {
+        for (int j = (int) path_->Size() - ((int) edges - (int) i); j >= 0; j -= (int) edges) {
             if (path_->operator [](j) != e) {
                 return false;
             }
